+++ NEWS.md
@@ -9,7 +9,7 @@ version 0.2.5.9000
 
 ## new features
 
-* New function `restartSpades`, which is still experimental. Its purpose is to be able to restart a simulation in the case of an error.
+* New function `restartSpades` and its associated `options(spades.recoveryMode = 1)`, the new default, which is still experimental. Its purpose is to be able to restart a simulation in the case of an error or interruption. 
 * Now gives better errors if modules are missing main .R file or if they are missing entirely
 * More silent tests
 * `mod` is now an active binding to `sim[[currentModule(sim)]]$.objects` and its parent environment is `emptyenv()`. This should cause no changes to users who use `mod$...`, but