+++ NEWS.md
@@ -10,7 +10,8 @@ version 0.2.5.9000
 
 ## new features
 
-* New function `outputObjectNames` will extract just the object names of all 
+* `.robustDigest` method for `simList` class objects now does only includes parameters that are listed within the module metadata, if `Cache` or `.robustDigest` is called within a module. This means that changes to parameter values in "other" modules will not affect the Caching of "the current" module.
+* New function `outputObjectNames` will extract just the object names of all `outputObjects` across modules
 * New function `restartSpades` and its associated `options(spades.recoveryMode = 1)`, the new default, which is still experimental. Its purpose is to be able to restart a simulation in the case of an error or interruption. 
 * Now gives better errors if modules are missing main .R file or if they are missing entirely 
 * More silent tests