+++ NEWS.md
@@ -1,6 +1,6 @@
 Known issues: https://github.com/PredictiveEcology/SpaDES.core/issues
 
-version 0.2.5.9000
+version 0.2.6
 =============
 
 ## dependencies